Commands should speak with client

+++ frontend/src/main/kotlin/commands/CreateLabelCommand.kt
@@ -6,10 +6,11 @@ import models.Label
 
 class CreateLabelCommand(private val newLabel: Label, private val controller: GroupViewController) : Command {
     private val client = TODOClient()
+    private var existingLabel = false
 
     override fun execute() {
         // If label already exists, don't create it
-        var existingLabel = controller.labels().any { label -> label.name == newLabel.name }
+        existingLabel = client.getLabels().any { label -> label.name == newLabel.name }
         if (existingLabel) return
 
         client.createLabel(newLabel)
@@ -17,6 +18,9 @@ class CreateLabelCommand(private val newLabel: Label, private val controller: Gr
     }
 
     override fun undo() {
+        // If label wasn't created because it was a duplicate, then don't delete it
+        if (existingLabel) return
+
         client.deleteLabel(newLabel)
         controller.reloadGroupView()
     }
